Huge performance bug fixed + other bugfixes BTW, some bugs are still exist (e.g. look scene6.test) and performance could be improved more

+++ camera.cpp
@@ -1,6 +1,6 @@
 #include "camera.h"
 
-Camera::Camera(glm::dvec3 _eye, glm::dvec3 _center, glm::dvec3 _up, double _fovx, double _fovy, int _width, int _height):
+Camera::Camera(glm::vec3 _eye, glm::vec3 _center, glm::vec3 _up, float _fovx, float _fovy, int _width, int _height):
     eye(_eye), center(_center), up(_up), fovx(_fovx), fovy(_fovy), width(_width), height(_height)
 {
 }
