+++ camera.h
@@ -8,21 +8,21 @@ class Scene;
 
 class Camera
 {
-    glm::dvec3 eye;
-    glm::dvec3 center;
-    glm::dvec3 up;
-    double fovx, fovy;
+    glm::vec3 eye;
+    glm::vec3 center;
+    glm::vec3 up;
+    float fovx, fovy;
     int width, height;
 public:
-    Camera(glm::dvec3 _eye, glm::dvec3 _center, glm::dvec3 _up, double _fovx, double _fovy, int _width, int _height);
+    Camera(glm::vec3 _eye, glm::vec3 _center, glm::vec3 _up, float _fovx, float _fovy, int _width, int _height);
     Camera();
-    glm::dvec3 getEye() const {return eye;}
-    glm::dvec3 getCenter() const {return center;}
-    glm::dvec3 getUp() const {return up;}
+    glm::vec3 getEye() const {return eye;}
+    glm::vec3 getCenter() const {return center;}
+    glm::vec3 getUp() const {return up;}
     int getWidth() const {return width;}
     int getHeight() const {return height;}
-    double getFovx() const {return fovx;}
-    double getFovy() const {return fovy;}
+    float getFovx() const {return fovx;}
+    float getFovy() const {return fovy;}
     friend std::pair<Camera *, Scene *> readfile(const char* filename);
 };
