+++ intersection.cpp
@@ -1,6 +1,6 @@
 #include "intersection.h"
 
-Intersection::Intersection(glm::dvec3 _normal, Ray _ray, double _t, Object* obj):
+Intersection::Intersection(glm::vec3 _normal, Ray _ray, float _t, Object* obj):
     normal(_normal), ray(_ray), t(_t), object(obj)
 {
 }
