+++ object.cpp
@@ -2,5 +2,5 @@
 
 Object::Object()
 {
-    mat.ambient = glm::dvec3(.2, .2, .2);
+    mat.ambient = glm::vec3(.2, .2, .2);
 }
