+++ readfile.h
@@ -14,7 +14,7 @@ using namespace std;
 class Camera;
 class Scene;
 
-void matransform (stack<glm::dmat4> &transfstack, GLdouble * values) ;
-void rightmultiply (const glm::dmat4 & M, stack<glm::dmat4> &transfstack) ;
-bool readvals (stringstream &s, const int numvals, GLdouble * values) ;
+void matransform (stack<glm::mat4> &transfstack, GLfloat * values) ;
+void rightmultiply (const glm::mat4 & M, stack<glm::mat4> &transfstack) ;
+bool readvals (stringstream &s, const int numvals, GLfloat * values) ;
 pair<Camera*, Scene*> readfile (const char * filename) ;
